setup/requirements.txt
pandas==1.3.0
bs4==0.0.1
lxml==4.6.3
html5lib==1.1
pillow